+++ kytos/core/interface.py
@@ -1,8 +1,10 @@
 """Module with main classes related to Interfaces."""
 import json
 import logging
+import operator
 from collections import OrderedDict
 from enum import IntEnum
+from functools import reduce
 from threading import Lock
 
 from pyof.v0x01.common.phy_port import Port as PortNo01
@@ -65,6 +67,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
     """Interface Class used to abstract the network interfaces."""
 
     status_funcs = OrderedDict()
+    status_reason_funcs = OrderedDict()
 
     # pylint: disable=too-many-arguments, too-many-public-methods
     def __init__(self, name, port_number, switch, address=None, state=None,
@@ -151,6 +154,24 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
     def register_status_func(cls, name: str, func):
         """Register status func given its name and a callable at setup time."""
         cls.status_funcs[name] = func
+
+    @classmethod
+    def register_status_reason_func(cls, name: str, func):
+        """Register status reason func given its name
+        and a callable at setup time."""
+        cls.status_reason_funcs[name] = func
+
+    @property
+    def status_reason(self):
+        """Return the reason behind the current status of the entity."""
+        return reduce(
+            operator.or_,
+            map(
+                lambda x: x(self),
+                self.status_reason_funcs.values()
+            ),
+            super().status_reason
+        )
 
     def set_available_tags(self, iterable):
         """Set a range of VLAN tags to be used by this Interface.
@@ -422,6 +443,7 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
              'lldp': True,
              'active': True,
              'enabled': False,
+             'status': 'DISABLED',
              'link': ""
             }
 
@@ -429,20 +451,24 @@ class Interface(GenericEntity):  # pylint: disable=too-many-instance-attributes
             dict: Dictionary filled with interface attributes.
 
         """
-        iface_dict = {'id': self.id,
-                      'name': self.name,
-                      'port_number': self.port_number,
-                      'mac': self.address,
-                      'switch': self.switch.dpid,
-                      'type': 'interface',
-                      'nni': self.nni,
-                      'uni': self.uni,
-                      'speed': self.speed,
-                      'metadata': self.metadata,
-                      'lldp': self.lldp,
-                      'active': self.is_active(),
-                      'enabled': self.is_enabled(),
-                      'link': self.link.id if self.link else ""}
+        iface_dict = {
+            'id': self.id,
+            'name': self.name,
+            'port_number': self.port_number,
+            'mac': self.address,
+            'switch': self.switch.dpid,
+            'type': 'interface',
+            'nni': self.nni,
+            'uni': self.uni,
+            'speed': self.speed,
+            'metadata': self.metadata,
+            'lldp': self.lldp,
+            'active': self.is_active(),
+            'enabled': self.is_enabled(),
+            'status': self.status.value,
+            'status_reason': sorted(self.status_reason),
+            'link': self.link.id if self.link else "",
+        }
         if self.stats:
             iface_dict['stats'] = self.stats.as_dict()
         return iface_dict
@@ -492,10 +518,20 @@ class UNI:
         return (self.user_tag == other.user_tag and
                 self.interface == other.interface)
 
+    def _is_reserved_valid_tag(self) -> bool:
+        """Check if TAG string is possible"""
+        reserved_tag = {"any", "untagged"}
+        if self.user_tag.value in reserved_tag:
+            return True
+        return False
+
     def is_valid(self):
         """Check if TAG is possible for this interface TAG pool."""
         if self.user_tag:
-            return self.interface.is_tag_available(self.user_tag)
+            if isinstance(self.user_tag.value, str):
+                return self._is_reserved_valid_tag()
+            if isinstance(self.user_tag.value, int):
+                return self.interface.is_tag_available(self.user_tag)
         return True
 
     def as_dict(self):