+++ kytos/core/switch.py
@@ -1,7 +1,9 @@
 """Module with main classes related to Switches."""
 import json
 import logging
+import operator
 from collections import OrderedDict
+from functools import reduce
 from threading import Lock
 
 from kytos.core.common import EntityStatus, GenericEntity
@@ -49,6 +51,7 @@ class Switch(GenericEntity):
     """
 
     status_funcs = OrderedDict()
+    status_reason_funcs = OrderedDict()
 
     def __init__(self, dpid, connection=None, features=None):
         """Contructor of switches have the below parameters.
@@ -137,6 +140,24 @@ class Switch(GenericEntity):
     def register_status_func(cls, name: str, func):
         """Register status func given its name and a callable at setup time."""
         cls.status_funcs[name] = func
+
+    @classmethod
+    def register_status_reason_func(cls, name: str, func):
+        """Register status reason func given its name
+        and a callable at setup time."""
+        cls.status_reason_funcs[name] = func
+
+    @property
+    def status_reason(self):
+        """Return the reason behind the current status of the entity."""
+        return reduce(
+            operator.or_,
+            map(
+                lambda x: x(self),
+                self.status_reason_funcs.values()
+            ),
+            super().status_reason
+        )
 
     def disable(self):
         """Disable this switch instance.
@@ -348,7 +369,8 @@ class Switch(GenericEntity):
                 'interfaces': {},
                 'metadata': {},
                 'active': True,
-                'enabled': False
+                'enabled': False,
+                'status': 'DISABLED',
                 }
 
         Returns:
@@ -361,22 +383,28 @@ class Switch(GenericEntity):
             port = self.connection.port
             connection = f"{address}:{port}"
 
-        return {'id': self.id,
-                'name': self.id,
-                'dpid': self.dpid,
-                'connection': connection,
-                'ofp_version': self.ofp_version,
-                'type': 'switch',
-                'manufacturer': self.description.get('manufacturer', ''),
-                'serial': self.description.get('serial', ''),
-                'hardware': self.description.get('hardware', ''),
-                'software': self.description.get('software'),
-                'data_path': self.description.get('data_path', ''),
-                'interfaces': {i.id: i.as_dict()
-                               for i in self.interfaces.values()},
-                'metadata': self.metadata,
-                'active': self.is_active(),
-                'enabled': self.is_enabled()}
+        return {
+            'id': self.id,
+            'name': self.id,
+            'dpid': self.dpid,
+            'connection': connection,
+            'ofp_version': self.ofp_version,
+            'type': 'switch',
+            'manufacturer': self.description.get('manufacturer', ''),
+            'serial': self.description.get('serial', ''),
+            'hardware': self.description.get('hardware', ''),
+            'software': self.description.get('software'),
+            'data_path': self.description.get('data_path', ''),
+            'interfaces': {
+                i.id: i.as_dict()
+                for i in self.interfaces.values()
+            },
+            'metadata': self.metadata,
+            'active': self.is_active(),
+            'enabled': self.is_enabled(),
+            'status': self.status.value,
+            'status_reason': sorted(self.status_reason),
+        }
 
     def as_json(self):
         """Return JSON with switch's attributes.