+++ tests/unit/test_core/test_interface.py
@@ -301,12 +301,26 @@ class TestInterface(unittest.TestCase):
         self.iface.enable()
         self.iface.activate()
         assert self.iface.is_active()
-        Interface.register_status_func("some_napp_some_func",
-                                       lambda iface: EntityStatus.DOWN)
+        Interface.register_status_func(
+            "some_napp_some_func",
+            lambda iface: EntityStatus.DOWN
+        )
+        Interface.register_status_reason_func(
+            "some_napp_some_func",
+            lambda iface: {'test_status'}
+        )
         assert self.iface.status == EntityStatus.DOWN
-        Interface.register_status_func("some_napp_some_func",
-                                       lambda iface: None)
+        assert self.iface.status_reason == {'test_status'}
+        Interface.register_status_func(
+            "some_napp_some_func",
+            lambda iface: None
+        )
+        Interface.register_status_reason_func(
+            "some_napp_some_func",
+            lambda iface: set()
+        )
         assert self.iface.status == EntityStatus.UP
+        assert self.iface.status_reason == set()
 
 
 class TestUNI(unittest.TestCase):
@@ -337,6 +351,17 @@ class TestUNI(unittest.TestCase):
 
         self.uni.user_tag = None
         self.assertTrue(self.uni.is_valid())
+
+    def test_is_reserved_valid_tag(self):
+        """Test _is_reserved_valid_tag method for string TAG."""
+        self.uni.user_tag = TAG(1, "untagged")
+        self.assertTrue(self.uni.is_valid())
+
+        self.uni.user_tag = TAG(1, "any")
+        self.assertTrue(self.uni.is_valid())
+
+        self.uni.user_tag = TAG(1, "2/4094")
+        self.assertFalse(self.uni.is_valid())
 
     def test_as_dict(self):
         """Test as_dict method."""