+++ tests/unit/test_core/test_switch.py
@@ -6,7 +6,6 @@ from unittest import TestCase
 from unittest.mock import MagicMock, Mock, patch
 
 from kytos.core import Controller
-from kytos.core.buffers import KytosBuffers
 from kytos.core.common import EntityStatus
 from kytos.core.config import KytosConfig
 from kytos.core.constants import FLOOD_TIMEOUT
@@ -28,9 +27,8 @@ class TestSwitch(TestCase):
 
         self.options = KytosConfig().options['daemon']
         self.controller = Controller(self.options)
-        self.controller.buffers = KytosBuffers()
         self.controller.log = Mock()
-
+        self.controller._buffers = MagicMock()
         self.switch = self.create_switch()
 
     @staticmethod
@@ -332,42 +330,30 @@ class TestSwitch(TestCase):
         self.assertIsNone(expected_ports_1)
         self.assertEqual(expected_ports_2, [1, 2, 3])
 
-    def test_as_dict(self):
-        """Test as_dict method."""
-        expected_dict = {'id': '00:00:00:00:00:00:00:01',
-                         'name': '00:00:00:00:00:00:00:01',
-                         'dpid': '00:00:00:00:00:00:00:01',
-                         'connection': 'addr:port',
-                         'ofp_version': '0x04',
-                         'type': 'switch',
-                         'manufacturer': '',
-                         'serial': '',
-                         'hardware': '',
-                         'software': None,
-                         'data_path': '',
-                         'interfaces': {},
-                         'metadata': {},
-                         'active': True,
-                         'enabled': True}
+    def test_as_dict_as_json(self):
+        """Test as_dict and as_json method."""
+        expected_dict = {
+            'id': '00:00:00:00:00:00:00:01',
+            'name': '00:00:00:00:00:00:00:01',
+            'dpid': '00:00:00:00:00:00:00:01',
+            'connection': 'addr:port',
+            'ofp_version': '0x04',
+            'type': 'switch',
+            'manufacturer': '',
+            'serial': '',
+            'hardware': '',
+            'software': None,
+            'data_path': '',
+            'interfaces': {},
+            'metadata': {},
+            'active': True,
+            'enabled': True,
+            'status': 'UP',
+            'status_reason': []
+        }
         self.assertEqual(self.switch.as_dict(), expected_dict)
 
-    def test_as_json(self):
-        """Test as_json method."""
-        expected_json = json.dumps({'id': '00:00:00:00:00:00:00:01',
-                                    'name': '00:00:00:00:00:00:00:01',
-                                    'dpid': '00:00:00:00:00:00:00:01',
-                                    'connection': 'addr:port',
-                                    'ofp_version': '0x04',
-                                    'type': 'switch',
-                                    'manufacturer': '',
-                                    'serial': '',
-                                    'hardware': '',
-                                    'software': None,
-                                    'data_path': '',
-                                    'interfaces': {},
-                                    'metadata': {},
-                                    'active': True,
-                                    'enabled': True})
+        expected_json = json.dumps(expected_dict)
 
         self.assertEqual(self.switch.as_json(), expected_json)
 
@@ -385,9 +371,23 @@ class TestSwitch(TestCase):
         self.switch.enable()
         self.switch.activate()
         assert self.switch.is_active()
-        Switch.register_status_func("some_napp_some_func",
-                                    lambda switch: EntityStatus.DOWN)
+        Switch.register_status_func(
+            "some_napp_some_func",
+            lambda switch: EntityStatus.DOWN
+        )
+        Switch.register_status_reason_func(
+            "some_napp_some_func",
+            lambda iface: {'test_status'}
+        )
         assert self.switch.status == EntityStatus.DOWN
-        Switch.register_status_func("some_napp_some_func",
-                                    lambda iface: None)
+        assert self.switch.status_reason == {'test_status'}
+        Switch.register_status_func(
+            "some_napp_some_func",
+            lambda iface: None
+        )
+        Switch.register_status_reason_func(
+            "some_napp_some_func",
+            lambda iface: set()
+        )
         assert self.switch.status == EntityStatus.UP
+        assert self.switch.status_reason == set()